export only React components in files to fix hot reloading
