+++ apps/fishing-map/features/datasets/upload/NewDataset.tsx
@@ -18,8 +18,8 @@ import NewPolygonDataset from 'features/datasets/upload/NewPolygonDataset'
 import NewTrackDataset from 'features/datasets/upload/NewTrackDataset'
 import { selectDataviewInstancesMerged } from 'features/dataviews/selectors/dataviews.resolvers.selectors'
 import UserGuideLink from 'features/help/UserGuideLink'
-import { getModalParent } from 'features/modals/Modals'
 import type { DatasetUploadStyle } from 'features/modals/modals.slice'
+import { getModalParent } from 'features/modals/modals.utils'
 import { selectIsGuestUser, selectIsUserExpired } from 'features/user/selectors/user.selectors'
 import { RegisterOrLoginToUpload } from 'features/workspace/user/UserSection/UserSection'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
