+++ apps/fishing-map/features/download/DownloadActivityModal.tsx
@@ -23,7 +23,7 @@ import {
 } from 'features/download/downloadActivity.slice'
 import DownloadActivityEnvironment from 'features/download/DownloadActivityEnvironment'
 import DownloadSurvey, { DISABLE_DOWNLOAD_SURVEY } from 'features/download/DownloadSurvey'
-import { getModalParent } from 'features/modals/Modals'
+import { getModalParent } from 'features/modals/modals.utils'
 
 import { HeatmapDownloadTab } from './downloadActivity.config'
 import DownloadActivityByVessel from './DownloadActivityByVessel'
