+++ apps/fishing-map/features/map/popups/categories/VesselEventsLayers.tsx
@@ -18,8 +18,8 @@ import { getDatasetSourceTranslated } from 'features/i18n/utils.datasets'
 import { selectVisibleResources } from 'features/resources/resources.selectors'
 import { DEFAULT_VESSEL_IDENTITY_ID } from 'features/vessel/vessel.config'
 import { getVesselProperty } from 'features/vessel/vessel.utils'
+import type { VesselPinOnClickCb } from 'features/vessel/vessel-pin.hooks'
 import VesselLink from 'features/vessel/VesselLink'
-import type { VesselPinOnClickCb } from 'features/vessel/VesselPin'
 import VesselPin from 'features/vessel/VesselPin'
 import { selectIsAnyVesselLocation } from 'router/routes.selectors'
 import { getEventDescription, getTimeLabels } from 'utils/events'
