+++ apps/fishing-map/features/modals/modals.utils.ts
@@ -1,0 +1,1 @@
+export const getModalParent = () => document.getElementById('app-layout-content') as HTMLElement
