+++ apps/fishing-map/features/reports/shared/new-report-modal/NewAreaReportModal.tsx
@@ -18,7 +18,7 @@ import { useAppDispatch } from 'features/app/app.hooks'
 import { selectWorkspaceWithCurrentState } from 'features/app/selectors/app.workspace.selectors'
 import { selectDatasetAreaDetail } from 'features/areas/areas.slice'
 import { selectPrivateDatasetsInWorkspace } from 'features/dataviews/selectors/dataviews.selectors'
-import { getModalParent } from 'features/modals/Modals'
+import { getModalParent } from 'features/modals/modals.utils'
 import { selectReportAreaIds } from 'features/reports/report-area/area-reports.selectors'
 import { createReportThunk, updateReportThunk } from 'features/reports/reports.slice'
 import { selectUserData } from 'features/user/selectors/user.selectors'
