+++ apps/fishing-map/features/reports/shared/vessels/ReportVesselsTable.tsx
@@ -26,8 +26,8 @@ import type {
 import { selectReportIsPinningVessels } from 'features/reports/tabs/activity/reports-activity.slice'
 import type { ReportActivityUnit } from 'features/reports/tabs/activity/reports-activity.types'
 import { selectUserData } from 'features/user/selectors/user.selectors'
+import type { VesselPinClickProps } from 'features/vessel/vessel-pin.hooks'
 import VesselLink from 'features/vessel/VesselLink'
-import type { VesselPinClickProps } from 'features/vessel/VesselPin'
 import VesselPin from 'features/vessel/VesselPin'
 import { selectWorkspaceStatus } from 'features/workspace/workspace.selectors'
 import { useReplaceQueryParams } from 'router/routes.hook'
