+++ apps/fishing-map/features/vessel/related-vessels/RelatedVessel.tsx
@@ -6,8 +6,8 @@ import type { Dataset, IdentityVessel } from '@globalfishingwatch/api-types'
 
 import { selectVesselDataset } from 'features/vessel/selectors/vessel.resources.selectors'
 import { getCurrentIdentityVessel } from 'features/vessel/vessel.utils'
+import type { VesselToResolve } from 'features/vessel/vessel-pin.hooks'
 import VesselLink from 'features/vessel/VesselLink'
-import type { VesselToResolve } from 'features/vessel/VesselPin'
 import VesselPin from 'features/vessel/VesselPin'
 import { selectIsWorkspaceVesselLocation } from 'router/routes.selectors'
 import { formatInfoField } from 'utils/info'
