+++ apps/fishing-map/features/vessel/vesselCorrection/VesselCorrectionModal.tsx
@@ -10,7 +10,7 @@ import { PATH_BASENAME, ROOT_DOM_ELEMENT } from 'data/config'
 import flags from 'data/flags'
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
 import I18nFlag from 'features/i18n/i18nFlag'
-import { getModalParent } from 'features/modals/Modals'
+import { getModalParent } from 'features/modals/modals.utils'
 import GFWOnly from 'features/user/GFWOnly'
 import { selectUserData } from 'features/user/selectors/user.selectors'
 import { selectVesselInfoData } from 'features/vessel/selectors/vessel.selectors'
