+++ apps/fishing-map/features/workspace/context-areas/ContextAreaLayerPanel.tsx
@@ -27,7 +27,7 @@ import { selectBasemapLabelsDataviewInstance } from 'features/dataviews/selector
 import { selectDebugOptions } from 'features/debug/debug.slice'
 import ContextLayerReportLink from 'features/map/popups/categories/ContextLayerReportLink'
 import { useContextInteractions } from 'features/map/popups/categories/ContextLayers.hooks'
-import { getModalParent } from 'features/modals/Modals'
+import { getModalParent } from 'features/modals/modals.utils'
 import GFWOnly from 'features/user/GFWOnly'
 import { selectIsGuestUser } from 'features/user/selectors/user.selectors'
 import {
