+++ apps/fishing-map/features/workspace/environmental/EnvironmentalLayerPanel.tsx
@@ -19,10 +19,9 @@ import { useActivityDataviewId } from 'features/map/map-layers.hooks'
 import { selectIsGFWUser } from 'features/user/selectors/user.selectors'
 import DatasetSchemaField from 'features/workspace/shared/DatasetSchemaField'
 import ExpandedContainer from 'features/workspace/shared/ExpandedContainer'
+import { isHistogramDataviewSupported } from 'features/workspace/shared/layer-filters.utils'
 import { useLayerPanelDataviewSort } from 'features/workspace/shared/layer-panel-sort.hook'
-import ActivityFilters, {
-  isHistogramDataviewSupported,
-} from 'features/workspace/shared/LayerFilters'
+import ActivityFilters from 'features/workspace/shared/LayerFilters'
 import { showSchemaFilter } from 'features/workspace/shared/LayerSchemaFilter'
 import MapLegend from 'features/workspace/shared/MapLegend'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
