+++ apps/fishing-map/router.tsx
@@ -1,26 +1,14 @@
-import { useMemo } from 'react'
-import { Provider } from 'react-redux'
 import { createRouter, Navigate } from '@tanstack/react-router'
 
 import { parseWorkspace, stringifyWorkspace } from '@globalfishingwatch/dataviews-client'
 
 // import { Spinner } from '@globalfishingwatch/ui-components'
 import { PATH_BASENAME } from 'data/config'
-import ErrorBoundaryUI from 'features/app/ErrorBoundaryUI'
-import { makeStore } from 'store'
+import { RouterErrorBoundary } from 'features/app/ErrorBoundaryRouter'
 import type { QueryParams } from 'types'
 
 import { setRouterRef } from './router/router-ref'
 import { routeTree } from './routeTree.gen'
-
-function RouterErrorBoundary({ error }: { error: Error }) {
-  const store = useMemo(() => makeStore(), [])
-  return (
-    <Provider store={store}>
-      <ErrorBoundaryUI error={error} />
-    </Provider>
-  )
-}
 
 const parseAppWorkspace = (searchStr: string): QueryParams => {
   return parseWorkspace(searchStr) as QueryParams
